move runnable entrypoints into examples/
Moves the demo/PoC scripts out of the repo root into a conventional
examples/ directory:

  anomaly_poc.py     -> examples/anomaly_detection.py
  simulation_run.py  -> examples/simulation.py
  twin_paradox.py    -> examples/twin_paradox.py
  sanity_harness.py  -> examples/sanity_harness.py
  calibration_harness.py -> examples/calibration_harness.py

tests/conftest.py now adds examples/ to sys.path so test modules can
still import them directly. tests/test_anomaly_poc.py is renamed to
tests/test_anomaly_detection.py to match.

Updated: README example-run commands,
scripts/check_to_packet_contract.py target paths,
tests/test_packet_contract_check.py expected target names,
tests/test_harness_regressions_v020.py filesystem lookup.

Repo root now holds only config (tg.yaml, pyproject.toml), docs, and
the temporal_gradient package — no runnable scripts. 166 tests still
passing.
