+++ examples/anomaly_detection.py
