+++ tests/conftest.py
@@ -2,5 +2,6 @@ import sys
 from pathlib import Path
 
 ROOT = Path(__file__).resolve().parents[1]
-if str(ROOT) not in sys.path:
-    sys.path.insert(0, str(ROOT))
+for path in (ROOT, ROOT / "examples"):
+    if str(path) not in sys.path:
+        sys.path.insert(0, str(path))
